Add files via upload
This is an incomplete version of an inventory management system I'm currently building. I'm currently searching for an API that will allow the system to use a device's camera to scan a barcode. Also, hardly any work has been done regarding the CSS and multiple pages have yet to be incorporated into the working system.